pyproject.toml
[tool.poetry]
name = "ragnarok-bro-api"
packages = [
    { include = "app" },
    { include = "web_scraping" }
]
version = "0.1.0"
description = ""
authors = ["Gabriel <gabrielsalves2k18@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
psycopg2 = "^2.9.9"
sqlalchemy = "^2.0.23"
selenium = "^4.16.0"
alembic = "^1.13.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
pytest = "^7.4.3"
httpx = "^0.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
